Added multicasting! This eliminates the need to enter an IP address manually, as potential receivers (peers) automatically announce themselves.
This new code needs to be cleaned.

+++ src/Transfer.java
@@ -14,11 +14,11 @@ public abstract class Transfer extends Thread
 	{
 		WAITING, REJECTED, TRANSFERRING, VERIFYING, FINISHED, FAILED
 	}
-	
+
 	/**
 	 * The file is sent in a series of chunks; here we define how big those are (in bytes).
 	 */
-	public final static int CHUNK_SIZE = (int) ( 1024 * 4 );
+	public final static int CHUNK_SIZE = ( 1024 * 4 );
 
 	/**
 	 * Which stage this transfer is in.
